+++ js/src/widget.js
@@ -4,7 +4,12 @@ export class DictWidget {
 
     static initialize(context) {
         $('table.dictwidget', context).each(function() {
-            new DictWidget($(this));
+            let elem = $(this);
+            if (window.yafowil_array !== undefined &&
+                window.yafowil_array.inside_template(elem)) {
+                return;
+            }
+            new DictWidget(elem);
         });
     }
 
@@ -149,5 +154,23 @@ export class DictWidget {
         let row = this.get_row(evt.currentTarget);
         row.insertAfter(row.next());
         this.reset_indices(row.parent());
+    }
+}
+
+//////////////////////////////////////////////////////////////////////////////
+// yafowil.widget.array integration
+//////////////////////////////////////////////////////////////////////////////
+
+function dict_on_array_add(inst, context) {
+    DictWidget.initialize(context);
+}
+
+export function register_array_subscribers() {
+    if (window.yafowil_array !== undefined) {
+        window.yafowil_array.on_array_event('on_add', dict_on_array_add);
+    } else if (yafowil.array !== undefined) {
+        $.extend(yafowil.array.hooks.add, {
+            dictwidget_binder: DictWidget.initialize
+        });
     }
 }