+++ CHANGES.rst
@@ -4,6 +4,10 @@ Changes
 2.0 (unreleased)
 ----------------
 
+- Extend JS by ``dict_on_array_add`` and ``register_array_subscribers``
+  functions to enable usage in ``yafowil.widget.array``.
+  [lenadax]
+
 - Add support for key and value datatypes.
   [rnix]